moved propertystore realted classes to settingsservice folder
