+++ Template10Library/Template10Library.csproj
@@ -143,8 +143,8 @@
     <Compile Include="Services\NavigationService\NavigatingEventArgs.cs" />
     <Compile Include="Services\NavigationService\NavigationService.cs" />
     <Compile Include="Services\SettingsService\Settings.cs" />
-    <Compile Include="Utils\PropertyMapping.cs" />
-    <Compile Include="Utils\PropertyStore.cs" />
+    <Compile Include="Services\SettingsService\PropertyMapping.cs" />
+    <Compile Include="Services\SettingsService\PropertyStore.cs" />
     <Compile Include="Utils\TypeUtil.cs" />
     <Content Include="Properties\Template10Library.rd.xml" />
   </ItemGroup>
