+++ Template10LibraryTests/Template10LibraryTests.csproj
@@ -100,7 +100,7 @@
     <Compile Include="UnitTestApp.xaml.cs">
       <DependentUpon>UnitTestApp.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Utils\JsonStoreTest.cs" />
+    <Compile Include="Services\SettingsService\JsonStoreTest.cs" />
     <Compile Include="Utils\TypeUtilTest.cs" />
   </ItemGroup>
   <ItemGroup>
